requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5
Jinja2==3.1.3
llm==0.13.1
llm-gpt4all==0.4
openai==1.25.0
prompt_toolkit==3.0.43
requests==2.31.0
tiktoken==0.6.0
unstructured==0.13.6
poppler-utils==0.1.0
python-slugify==8.0.4
requests==2.31.0
